deleted unused func, new folder struct, new file struct, Feed struct changed
