+++ RssOnVIPER/Base.lproj/Main.storyboard
@@ -73,7 +73,7 @@
         <!--Favorite-->
         <scene sceneID="DAT-Dl-Eei">
             <objects>
-                <viewController id="Qnw-Ku-K17" customClass="FeedFavoriteListViewController" customModule="RssOnVIPER" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="Qnw-Ku-K17" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="X9v-Qk-6iJ">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
