--- RssOnVIPER/FeedFavoriteList/FeedFavoriteListRouter.swift
@@ -1,14 +1,0 @@
-//
-//  FeedFavoriteListRouter.swift
-//  RssOnVIPER
-//
-//  Created by Anton Tsykanov on 11.09.2018.
-//  Copyright © 2018 Anton Tsykanov. All rights reserved.
-//
-
-import Foundation
-
-protocol FeedFavoriteListRouter {
-    
-    func openDetails(item: FeedListFavoriteEntity)
-}
