--- RssOnVIPER/PostList/FeedListRouter.swift
@@ -1,15 +1,0 @@
-//
-//  FeedListRouter.swift
-//  RssOnVIPER
-//
-//  Created by Anton Tsykanov on 10.09.2018.
-//  Copyright © 2018 Anton Tsykanov. All rights reserved.
-//
-
-import Foundation
-
-
-protocol FeedListRouterProtocol {
-    
-    func openDetails(item: FeedListEntity)
-}
