+++ RssOnVIPER/UIModules/FeedListModule/Entities/FeedListEntity.swift
@@ -1,0 +1,16 @@
+//
+//  FeedListEntity.swift
+//  RssOnVIPER
+//
+//  Created by Anton Tsykanov on 10.09.2018.
+//  Copyright © 2018 Anton Tsykanov. All rights reserved.
+//
+
+import Foundation
+
+struct FeedVM {
+    let imageUrl: String
+    let imagePlaceholderName: String
+    let title: String
+    let details: String
+}
