+++ RssOnVIPER/UIModules/FeedListModule/Interactor/FeedListInteractor.swift
@@ -1,0 +1,17 @@
+//
+//  FeedListInteractor.swift
+//  RssOnVIPER
+//
+//  Created by Anton Tsykanov on 10.09.2018.
+//  Copyright © 2018 Anton Tsykanov. All rights reserved.
+//
+
+import Foundation
+
+class FeedListInteractor : FeedListInteractorProtocol {
+    
+    func loadFeed() -> [FeedVM] {
+        return Array()
+    }
+    
+}
