+++ RssOnVIPER/UIModules/FeedListModule/Router/FeedListRouter.swift
@@ -1,0 +1,12 @@
+//
+//  FeedListRouter.swift
+//  RssOnVIPER
+//
+//  Created by Anton Tsykanov on 10.09.2018.
+//  Copyright © 2018 Anton Tsykanov. All rights reserved.
+//
+
+import Foundation
+
+
+
